+++ packages/scaffold-config/test/unit/detect.spec.ts
@@ -171,7 +171,7 @@ describe('detectFramework', () => {
     })
   })
 
-  ;['13.0.0', '14.0.0'].forEach((v) => {
+  ;['17.2.0', '18.2.0'].forEach((v) => {
     it(`Angular CLI v${v}`, async () => {
       const projectPath = await scaffoldMigrationProject('angular-cli-unconfigured')