+++ system-tests/projects/angular-cli-configured/src/app/app.component.cy.ts
@@ -2,5 +2,5 @@ import { AppComponent } from './app.component'
 
 it('should pass', () => {
   cy.mount(AppComponent)
-  cy.contains('angular-cli-app')
+  cy.contains('angular-cli-configured')
 })