+++ system-tests/projects/angular-cli-unconfigured/src/app/app.component.ts
@@ -1,10 +1,13 @@
 import { Component } from '@angular/core'
+import { RouterOutlet } from '@angular/router'
 
 @Component({
   selector: 'app-root',
+  standalone: true,
+  imports: [RouterOutlet],
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
+  styleUrl: './app.component.css',
 })
 export class AppComponent {
-  title = 'angular-cli-app'
+  title = 'angular-cli-unconfigured'
 }